Support a better set of locales
The formatter now supports a full set of Unicode-cldr locales. It's packed down to all "root level" languages and any sub-dialects that deviate from their "parent"language. Calling the formatter with a dialect that isn't there will default it to the language so it should do the right thing for the entire set of cldr locales.

fixes #1